requirements.txt
numpy==1.26.4
scipy==1.11.4
pandas==2.2.2
matplotlib==3.8.2
pyyaml==6.0.2
pydantic==2.8.2
